fix(speed): speed visualization

+++ lib/johnson_motor/johnson_motor.cpp
@@ -98,7 +98,7 @@ double JohnsonMotor::read_speed(void) {
 
 /// @brief Actuates the PID controller and updates the speed of the motor
 /// @param delta Time in seconds since the last update
-void JohnsonMotor::update_speed(uint32_t delta) {
+void JohnsonMotor::update_speed(double delta) {
     this->speed = ((this->encoder_count - this->prev_encoder_count) * 60) / (177.6 * delta);
     this->prev_encoder_count = this->encoder_count;
